+++ src/app/app.config.ts
@@ -1,8 +1,10 @@
+import { OVERLAY_DEFAULT_CONFIG } from '@angular/cdk/overlay';
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { ApplicationConfig } from '@angular/core';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { NoPreloading, provideRouter, withPreloading } from '@angular/router';
 import { provideHotToastConfig } from '@ngxpert/hot-toast';
+import { provideIxIcons } from '@pyrophire/ix-libs';
 import { routes } from './app.routes';
 import { httpInterceptorProviders } from './interceptors';
 
@@ -12,6 +14,13 @@ export const appConfig: ApplicationConfig = {
         provideAnimations(),
         provideHttpClient(withInterceptorsFromDi()),
         httpInterceptorProviders,
-        provideHotToastConfig()
+        provideHotToastConfig(),
+        provideIxIcons(),
+        {
+            provide: OVERLAY_DEFAULT_CONFIG,
+            useValue: {
+                usePopover: false
+            }
+        }
     ]
 };