+++ src/main.ts
@@ -6,7 +6,7 @@ import { appConfig } from './app/app.config';
 
 bootstrapApplication(AppComponent, {
     ...appConfig,
-    providers: [provideZoneChangeDetection(),...(appConfig.providers || [])]
+    providers: [provideZoneChangeDetection(), ...(appConfig.providers || [])]
 }).catch((err) => console.error(err));
 
 // Globally silence console logs in production while keeping warnings/errors